Keyword
Listens for preprogrammed keywords and responds with appropriate information when needed. Example being when "discord" is mentioned in chat it can return a link to the discord.

User can update in chat via command to update link, cooldown, and other attributes that are saved to file in Chatbot directory within Documents.